==> src/3_Split_Put_and_Call_data.Rmd <==
---
title: "Modify_dfs"
author: "Caterina Stifano"
date: "30/4/2020"
output: html_document
---

```{r}
#clan the workspace
rm(list = ls())
```

```{r}
#packages
library(tidyverse)
library(fOptions)
library(ggplot2)
library(roll)
```

#LOADING
```{r}
#Loading the excel file
raw_data <- read.csv("../data/rawdata.csv")
```

#ORDER DATA BY DATE
```{r}
raw_data = raw_data %>% mutate(date=as.Date(date))
raw_data = raw_data[order(as.Date(raw_data$date, format = "%Y/%m/%d")),]
```

#remove non needed variables
```{r}
raw_data$bid <- NULL
raw_data$ask <- NULL
raw_data$BS5.strike <- NULL
raw_data$BS30.strike <- NULL
raw_data$BS60.strike <- NULL
raw_data$BS90.strike <- NULL
raw_data$BS120.strike <- NULL
raw_data$BSgarch.strike <- NULL
raw_data$lag_close <- NULL
```

#remove non representative observations
```{r}
#subset call
call <- subset(raw_data, cpflag == "C")

#1) price option that is traded above 10 basis points
call <- subset(call, mid > 0.1)

#2) keep only the call options which have maturity in 15 days or 2 years (252 is used as the number of trading days in one year)
call['d_maturity'] = call$maturity/252
call <- subset(call, d_maturity > 15/252 & d_maturity < 2)

#3)keep only the call option which are not extremely ITM or OTM therefore we keep only the moneyness between 0.5 and 1.55
call <- subset(call, moneyness > 0.5 & moneyness < 1.55 )

#4)keep only the options that don't violate the lower bound conditon see page 262
call['lb_c'] <- call$close - (call$strike*exp(-call$interest*call$d_maturity))
call <- subset(call, mid >= lb_c)
```

```{r}
#subset put
put <- subset(raw_data, cpflag == "P")

#1) price option that is traded above 10 basis points
put <- subset(put, mid > 0.1)

#2) keep only the put options which have maturity in 15 days or 2 years (252 is used as the number of trading days in one year)
put['d_maturity'] = put$maturity/252
put <- subset(put, d_maturity > 15/252 & d_maturity < 2)

#3)keep only the put option which are not extremely ITM or OTM therefore we keep only the moneyness between 0.5 and 1.55
put <- subset(put, moneyness > 0.5 & moneyness < 1.55 )

#4)keep only the options that don't violate the lower bound conditon see page 262
put['lb_p'] <- (put$strike*exp(-put$interest*put$d_maturity)) - put$close 
put <- subset(put, mid >= lb_p)
```

#merge put and call together
```{r}
call$lb_c <- NULL
put$lb_p <- NULL
call$d_maturity <- NULL
put$d_maturity <- NULL
pdata <- rbind(call, put)

tot = nrow(call) + nrow(put)
```

##EXPORT THE DATASET
```{r}
write.csv(pdata, "../data/plot_data.csv")
write.csv(call, "../data/call.csv")
write.csv(put, "../data/put.csv")
```


#Analysis
```{r}
#see which are the most common strike for call and put

#call
c_l <- rle(sort(call$strike))
call_strikes <- data.frame(strikes=c_l$values, number=c_l$lengths)

#put
p_l <- rle(sort(put$strike))
put_strikes <- data.frame(strikes=p_l$values, number=p_l$lengths)

```
